+++ deps.js
@@ -1,0 +1,6 @@
+module.exports = {
+  React: require('react'),
+  ReactDOM: require('react-dom'),
+  lodash: require('lodash'),
+  hogan: require('hogan.js')
+};