+++ scripts/build.sh
@@ -10,15 +10,8 @@ license="/*! instantsearch.js ${VERSION:-UNRELEASED} | © Algolia Inc. and other
 
 bundle='instantsearch'
 
-printf "\n\nBuild"
-
-printf "\n\nBuild: webpack\n"
 webpack
 
-printf "\n\nBuild: minify\n"
-cat dist/$bundle.js | uglifyjs -c warnings=false -m > dist/$bundle.min.js
-
-printf "\n\nBuild: CSS\n"
 for source in "$ROOT"/themes/[^_]*.sass; do
   base=`basename "$source" .sass`
   echo "$license" > dist/themes/$base.css
@@ -27,10 +20,9 @@ for source in "$ROOT"/themes/[^_]*.sass; do
   cleancss dist/themes/$base.css > dist/themes/$base.min.css
 done
 
-printf "\n\nBuild: prepend license\n"
-printf "$license" | cat - dist/"$bundle".js > /tmp/out && mv /tmp/out dist/"$bundle".js
-printf "$license" | cat - dist/"$bundle".min.js > /tmp/out && mv /tmp/out dist/"$bundle".min.js
+printf "$license" | cat - dist/${bundle}.js > /tmp/out && mv /tmp/out dist/${bundle}.js
+cd dist
+uglifyjs ${bundle}.js --source-map ${bundle}.min.map --preamble "$license" -c warnings=false -m -o ${bundle}.min.js
+cd ..
 
-printf "\n\nBuild: filesize\n"
-
-printf "=> $bundle.min.js gzipped will weight `cat dist/$bundle.min.js | gzip -9 | wc -c | pretty-bytes`\n"
+printf "=> ${bundle}.min.js gzipped will weight `cat dist/${bundle}.min.js | gzip -9 | wc -c | pretty-bytes`\n"