+++ scripts/validate-commit-msgs.sh
@@ -1,5 +1,14 @@
 #!/usr/bin/env bash
 
+set -e # exit when error
+
+[ -z $TRAVIS_PULL_REQUEST ] && TRAVIS_PULL_REQUEST="false"
+
+if [ $TRAVIS_PULL_REQUEST == "false" ]; then
+  echo "No need to validate commit message when not in a pull request"
+  exit 0
+fi
+
 # Checks the commits msgs in the range of commits travis is testing.
 # Based heavily on
 # https://raw.githubusercontent.com/angular/angular.js/291d7c467fba51a9cb89cbeee62202d51fe64b09/validate-commit-msg.js