+++ widgets/index-selector/index-selector.js
@@ -2,6 +2,7 @@ var React = require('react');
 var ReactDOM = require('react-dom');
 
 var findIndex = require('lodash/array/findIndex');
+var map = require('lodash/collection/map');
 var utils = require('../../lib/utils.js');
 var bem = utils.bemHelper('ais-index-selector');
 var cx = require('classnames');
@@ -9,13 +10,13 @@ var autoHideContainer = require('../../decorators/autoHideContainer');
 
 /**
  * Instantiate a dropdown element to choose the current targeted index
- * @param  {String|DOMElement} options.container CSS Selector or DOMElement to insert the widget
+ * @param  {string|DOMElement} options.container CSS Selector or DOMElement to insert the widget
  * @param  {Array} options.indices Array of objects defining the different indices to choose from.
- * @param  {String} options.indices[0].name Name of the index to target
- * @param  {String} options.indices[0].label Label displayed in the dropdown
+ * @param  {string} options.indices[0].name Name of the index to target
+ * @param  {string} options.indices[0].label Label displayed in the dropdown
  * @param  {Object} [options.cssClasses] CSS classes to be added
- * @param  {String} [options.cssClasses.root] CSS classes added to the parent <select>
- * @param  {String} [options.cssClasses.item] CSS classes added to each <option>
+ * @param  {string} [options.cssClasses.root] CSS classes added to the parent <select>
+ * @param  {string} [options.cssClasses.item] CSS classes added to each <option>
  * @param  {boolean} [hideContainerWhenNoResults=false] Hide the container when no results match
  * @return {Object}
  */
@@ -27,17 +28,21 @@ function indexSelector({
   }) {
   var containerNode = utils.getContainerNode(container);
 
-  var usage = 'Usage: indexSelector({container, indices[, cssClasses.{select,option}, hideContainerWhenNoResults]})';
+  var usage = 'Usage: indexSelector({container, indices[, cssClasses.{root,item}, hideContainerWhenNoResults]})';
   if (!container || !indices) {
     throw new Error(usage);
   }
+
+  var selectorOptions = map(indices, function(index) {
+    return {label: index.label, value: index.name};
+  });
 
   return {
     init: function(state, helper) {
       var currentIndex = helper.getIndex();
       var isIndexInList = findIndex(indices, {name: currentIndex}) !== -1;
       if (!isIndexInList) {
-        throw new Error('[stats]: Index ' + currentIndex + ' not present in `indices`');
+        throw new Error('[indexSelector]: Index ' + currentIndex + ' not present in `indices`');
       }
     },
 
@@ -50,20 +55,20 @@ function indexSelector({
       let currentIndex = helper.getIndex();
       let hasResults = results.hits.length > 0;
       let setIndex = this.setIndex.bind(this, helper);
-      var IndexSelector = autoHideContainer(require('../../components/IndexSelector'));
+      var Selector = autoHideContainer(require('../../components/Selector'));
 
       cssClasses = {
         root: cx(bem(null), cssClasses.root),
         item: cx(bem('item'), cssClasses.item)
       };
       ReactDOM.render(
-        <IndexSelector
+        <Selector
           cssClasses={cssClasses}
-          currentIndex={currentIndex}
+          currentValue={currentIndex}
           hasResults={hasResults}
           hideContainerWhenNoResults={hideContainerWhenNoResults}
-          indices={indices}
-          setIndex={setIndex}
+          options={selectorOptions}
+          setValue={setIndex}
         />,
         containerNode
       );