+++ widgets/price-ranges/__tests__/price-ranges-test.js
@@ -21,7 +21,7 @@ describe('priceRanges()', () => {
   var autoHideContainer;
   var headerFooter;
 
-  jsdom();
+  jsdom({useEach: true});
 
   beforeEach(() => {
     ReactDOM = {render: sinon.spy()};
@@ -61,13 +61,15 @@ describe('priceRanges()', () => {
       };
 
       props = {
+        createURL: sinon.spy(),
         cssClasses: {
+          active: 'ais-price-ranges--range__active',
           body: 'ais-price-ranges--body',
           button: 'ais-price-ranges--button',
           footer: 'ais-price-ranges--footer',
           header: 'ais-price-ranges--header',
           input: 'ais-price-ranges--input',
-          form: 'ais-price-ranges--input-group',
+          form: 'ais-price-ranges--form',
           range: 'ais-price-ranges--range',
           root: 'ais-price-ranges'
         },